before merge with testing libraries from mike
