+++ src/constants/actionTypes.js
@@ -1,0 +1,1 @@
+// export const ACTION_DESCRIPTION = "ACTION_DESCRIPTION";
